+++ extremum-shared-models/src/main/java/io/extremum/sharedmodels/dto/Pagination.java
@@ -1,5 +1,6 @@
 package io.extremum.sharedmodels.dto;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -21,5 +22,27 @@ public class Pagination {
                 .count(total)
                 .total((long) total)
                 .build();
+    }
+
+    public Pagination() {
+    }
+
+    public Pagination(int offset, int count, Long total, ZonedDateTime since, ZonedDateTime until) {
+        this.offset = offset;
+        this.count = count;
+        this.total = total;
+        this.since = since;
+        this.until = until;
+    }
+
+    @Override
+    public String toString() {
+        return "Pagination{" +
+                "offset=" + offset +
+                ", count=" + count +
+                ", total=" + total +
+                ", since=" + since +
+                ", until=" + until +
+                '}';
     }
 }